requirements.txt
# GOATGuard Server — dependencias completas
# API + Pipeline + Detection

# API Framework
fastapi==0.135.1
uvicorn==0.42.0
pydantic>=2.0

# Base de datos
sqlalchemy>=2.0
psycopg2-binary==2.9.11

# Autenticación
bcrypt==5.0.0
PyJWT==2.12.1

# Configuración
pyyaml>=6.0

# Captura y análisis de red (pipeline)
scapy>=2.5
pythonping>=1.1
mac-vendor-lookup>=0.1

# WebSocket
websockets>=11.0

# TOTP / 2FA [RF-13]
pyotp==2.9.0
qrcode[pil]==8.2
cryptography==45.0.3

# Migraciones de BD
alembic>=1.13.0

# Validación HIBP + testing HTTP
httpx>=0.27.0

# Rate limiting
slowapi>=0.1.9
